Turned prefs into an enumeration for better programming.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@3092 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDentBusiness/Data Interface/ReplicationServers.cs
@@ -151,7 +151,7 @@ namespace OpenDentBusiness{
 		/*
 		private static long GetNumComputers() {
 			if(RemotingClient.RemotingRole==RemotingRole.ClientWeb) {
-				return Meth.GetInt(MethodBase.GetCurrentMethod());
+				return Meth.GetLong(MethodBase.GetCurrentMethod());
 			}
 			string command="SELECT COUNT(*) FROM computer";
 			DataTable table=Db.GetTable(command);
@@ -160,7 +160,7 @@ namespace OpenDentBusiness{
 
 		private static long GetComputerNumForName(string computerName) {
 			if(RemotingClient.RemotingRole==RemotingRole.ClientWeb) {
-				return Meth.GetInt(MethodBase.GetCurrentMethod(),computerName);
+				return Meth.GetLong(MethodBase.GetCurrentMethod(),computerName);
 			}
 			string command="SELECT COUNT(*) FROM computer "+
 				"WHERE ComputerNum <=(SELECT ComputerNum FROM computer AS temp WHERE CompName "+
